Fix todos

+++ webapps/smsc_cc/java/ru/novosoft/smsc/admin/filesystem/FileSystemSingleHA.java
@@ -2,7 +2,6 @@ package ru.novosoft.smsc.admin.filesystem;
 
 import ru.novosoft.smsc.admin.AdminException;
 import ru.novosoft.smsc.util.FileUtils;
-import ru.novosoft.smsc.util.Functions;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -16,8 +15,8 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public OutputStream getOutputStream(File file) throws AdminException {
-    assert file != null : "Some arguments are null";        // todo почему assert? Это же публичный метод. Тет правильнее IllegalArgument кидать или NullPointer
-                                                            // todo assert надо кидать в приватных методах.  
+    if (file == null)
+      throw new NullPointerException("File arg is null");
     try {
       return new FileOutputStream(file);
     } catch (IOException e) {
@@ -27,7 +26,9 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void rename(File file, File toFile) throws AdminException {
-    assert file != null && toFile != null : "Some arguments are null";
+    if (file == null || toFile == null)
+      throw new NullPointerException("File arg is null");
+
     if (!file.renameTo(toFile)) {
       String err = "Can't rename file '" + file.getAbsolutePath() + "' to '" + toFile.getAbsolutePath() + '\'';
       throw new FileSystemException("io_error",err);
@@ -36,7 +37,8 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void copy(File file, File toFile) throws AdminException {
-    assert file != null && toFile != null : "Some arguments are null";
+    if (file == null || toFile == null)
+      throw new NullPointerException("File arg is null");
     try {
       FileUtils.copyFileTo(file, toFile);
     } catch (IOException e) {
@@ -46,7 +48,8 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void delete(File file) throws AdminException {
-    assert file != null : "Some arguments are null";
+    if (file == null)
+      throw new NullPointerException("File arg is null");
     if (!file.delete() && file.exists()) {
       String err = "Can't remove file '" + file.getAbsolutePath() + '\'';
       throw new FileSystemException("io_error",err);
@@ -56,7 +59,8 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public void mkdirs(File file) throws AdminException {
-    assert file != null : "Some arguments are null";
+    if (file == null)
+      throw new NullPointerException("File arg is null");
     if (!file.mkdirs() && !file.exists()) {
       String err = "Can't create dirs for '" + file.getAbsolutePath() + '\'';
       throw new FileSystemException("io_error",err);
@@ -65,7 +69,8 @@ class FileSystemSingleHA extends FileSystem {
 
   @Override
   public boolean exists(File file) throws AdminException {
-    assert file != null : "Some arguments are null";
+    if (file == null)
+      throw new NullPointerException("File arg is null");
     return file.exists();
   }
 }
